+++ R/trace.R
@@ -5,6 +5,8 @@
 #' evaluation, the call stack in R is actually a tree, which the
 #' `summary()` method of this object will reveal.
 #'
+#' `trace_length()` returns the number of frames in a backtrace.
+#'
 #' @param top The first frame environment to be included in the
 #'   backtrace. This becomes the top of the backtrace tree and
 #'   represents the oldest call in the backtrace.
@@ -411,8 +413,11 @@ summary.rlang_trace <- function(object,
   invisible(object)
 }
 
-trace_length <- function(x) {
-  length(x$calls)
+#' @rdname trace_back
+#' @param trace A backtrace created by `trace_back()`.
+#' @export
+trace_length <- function(trace) {
+  length(trace$calls)
 }
 
 trace_subset <- function(x, i) {
@@ -506,7 +511,7 @@ trace_trim_env <- function(x, to = NULL) {
   start <- last(which(is_top)) + 1
   end <- length(x$envs)
 
-  trace_subset(x, start:end)
+  trace_subset(x, seq2(start, end))
 }
 
 set_trace_skipped <- function(trace, id, n) {