+++ NEWS.md
@@ -1,5 +1,10 @@
 
-# rlang 0.4.0.9000
+# rlang (development)
+
+* `trace_length()` returns the number of frames in a backtrace.
+
+* Added internal utility `cnd_entrace()` to add a backtrace to a
+  condition.
 
 * `rlang::last_error()` backtraces are no longer displayed in red.